change chain id

+++ AElf.Common/HexExtensions.cs
@@ -10,12 +10,12 @@ namespace AElf.Common
             return BitConverter.GetBytes(n).ToHex();
         }
 
-        public static string DumpBase58(this long n)
+        public static string DumpBase58(this int n)
         {
             return BitConverter.GetBytes(n).ToPlainBase58();
         }
 
-        public static byte[] DumpByteArray(this long n)
+        public static byte[] DumpByteArray(this int n)
         {
             return BitConverter.GetBytes(n);
         }
